IQ boost by considering both 1 part 1 plane block paths
The 1 partition 1 plane mode runs twice, once with a fast path
that considers a reduced set of block candidates and then again
with the full set. The first pass runs with a penalty to the
error threshold (only valid if it exceeds the error threshold
by a large margin).

This patch changes the logic to consider the first pass block
as a candidate for the second pass error metric. This gives a
small, but consistent, improvement to PSNR and is basically
free because we've computed it already.