Significantly improve the quality of -fastest
Improve the quality of -fastest by an average of 1dB, albeit at
around 0.5x the performance. With this change -fastest is ~2x
the speed of -fast.

The goal of this change was to improve predictability of the codec
and avoid the steep quality-to-perf drop-off for effort settings
under 10, where the loss in quality was much faster than the gain
in performance compared to effort settings above 10.